+++ clientai/agent/utils/validation.py
@@ -17,10 +17,10 @@ def validate_step(step: Step) -> None:
     Raises:
         StepError: If the step has no function or the function is not callable.
     """
-    if step.func is None:
+    if step.func is None:  # pragma: no cover
         raise StepError(f"Step {step.name} has no function")
 
-    if not callable(step.func):
+    if not callable(step.func):  # pragma: no cover
         raise StepError(f"Step {step.name} function is not callable")
 
 
@@ -36,10 +36,10 @@ def validate_tool(tool: Tool) -> None:
     Raises:
         ToolError: If the tool has no function or the function is not callable.
     """
-    if tool.func is None:
+    if tool.func is None:  # pragma: no cover
         raise ToolError(f"Tool {tool.name} has no function")
 
-    if not callable(tool.func):
+    if not callable(tool.func):  # pragma: no cover
         raise ToolError(f"Tool {tool.name} function is not callable")
 
 
@@ -68,5 +68,5 @@ def validate_callable(func: Callable) -> None:
         validate_callable(invalid_func)  # Raises ValueError
         ```
     """
-    if not callable(func):
+    if not callable(func):  # pragma: no cover
         raise ValueError(f"Object {func} is not callable")
