+++ tests/ollama/test_manager.py
@@ -90,7 +90,7 @@ def mock_platform_info():
                 if config.gpu_memory_fraction is not None:
                     env["CUDA_MEM_FRACTION"] = str(config.gpu_memory_fraction)
 
-            elif platform_info.gpu_vendor == GPUVendor.AMD:
+            elif platform_info.gpu_vendor == GPUVendor.AMD:  # pragma: no cover
                 if config.gpu_devices is not None:
                     devices = (
                         config.gpu_devices
@@ -110,9 +110,9 @@ def mock_platform_info():
         def get_server_command(config: Any) -> list[str]:
             """Get platform-specific server command."""
             base_cmd = ["ollama", "serve"]
-            if config.host != "127.0.0.1":
+            if config.host != "127.0.0.1":  # pragma: no cover
                 base_cmd.extend(["--host", config.host])
-            if config.port != 11434:
+            if config.port != 11434:  # pragma: no cover
                 base_cmd.extend(["--port", str(config.port)])
             return base_cmd
 
@@ -310,7 +310,7 @@ def test_gpu_specific_environment(gpu_vendor, config, expected_env):
                 if cfg.gpu_memory_fraction is not None:
                     env["CUDA_MEM_FRACTION"] = str(cfg.gpu_memory_fraction)
 
-            elif gpu_vendor == GPUVendor.AMD:
+            elif gpu_vendor == GPUVendor.AMD:  # pragma: no cover
                 if cfg.gpu_devices is not None:
                     devices = (
                         cfg.gpu_devices
@@ -414,7 +414,7 @@ def test_custom_host_port(mock_subprocess, mock_http_client):
 
     try:
         manager.start()
-    except ServerTimeoutError:
+    except ServerTimeoutError:  # pragma: no cover
         pass
 
     mock_http_client.assert_called_with("localhost", 11435, timeout=5)
